+++ lettings/views.py
@@ -1,3 +1,16 @@
 from django.shortcuts import render
+from lettings.models import Letting
 
-# Create your views here.
+def index(request):
+    lettings_list = Letting.objects.all()
+    context = {'lettings_list': lettings_list}
+    return render(request, 'lettings/index.html', context)
+
+
+def letting(request, letting_id):
+    letting = Letting.objects.get(id=letting_id)
+    context = {
+        'title': letting.title,
+        'address': letting.address,
+    }
+    return render(request, 'lettings/letting.html', context)
