+++ oc_lettings_site/urls.py
@@ -1,13 +1,11 @@
 from django.contrib import admin
-from django.urls import path
+from django.urls import path, include
 
-from . import views
+from .views import home
 
 urlpatterns = [
-    path('', views.index, name='index'),
-    path('lettings/', views.lettings_index, name='lettings_index'),
-    path('lettings/<int:letting_id>/', views.letting, name='letting'),
-    path('profiles/', views.profiles_index, name='profiles_index'),
-    path('profiles/<str:username>/', views.profile, name='profile'),
+    path('', home, name='home'),
+    path('', include('lettings.urls')),
+    path('', include('profiles.urls')),
     path('admin/', admin.site.urls),
 ]
