+++ templates/base.html
@@ -22,7 +22,7 @@
                     <!-- Navbar-->
                     <nav class="navbar  navbar-expand-lg bg-white navbar-light">
                         <div class="container">
-                            <a class="navbar-brand" href="{% url 'index'%}"><img class="img-responsive" src="{% static 'assets/img/logo.png' %}" width="70px" height="70px" alt="Logo Orange County Lettings"/></a>
+                            <a class="navbar-brand" href="{% url 'home'%}"><img class="img-responsive" src="{% static 'assets/img/logo.png' %}" width="70px" height="70px" alt="Logo Orange County Lettings"/></a>
                             <div>
                                 <a class="btn fw-500 ms-lg-4 btn-primary" href="{% url 'profiles_index' %}">
                                         Profiles
