+++ templates/lettings/index.html
@@ -33,7 +33,7 @@
 
 <div class="container px-5 py-5 text-center">
     <div class="justify-content-center">
-        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'index' %}">
+        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'home' %}">
             Home
         </a>
         <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'profiles_index' %}">
