+++ templates/lettings/letting.html
@@ -29,7 +29,7 @@
         	<i class="ms-2" data-feather="arrow-right"></i>
             Back
         </a>
-        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'index' %}">
+        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'home' %}">
             Home
         </a>
         <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'profiles_index' %}">
