+++ templates/profiles/profile.html
@@ -28,7 +28,7 @@
         	<i class="ms-2" data-feather="arrow-left"></i>
             Back
         </a>
-        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'index' %}">
+        <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'home' %}">
             Home
         </a>
         <a class="btn fw-500 ms-lg-4 btn-primary px-10" href="{% url 'lettings_index' %}">
